\name{lpls-package}
\alias{lpls-package}
\docType{package}
\title{
Lpls data exploration and regression
}
\description{
A package for performing endo or exo lpls regression.
}
\details{
\tabular{ll}{
Package: \tab lpls\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2016-07-18\cr
License: \tab GPL-2\cr
}
The main function for fitting an lpls-object is the function \code{lpls}. Cross-validation can be run from \code{lplsCV}.
}
\author{
Solve Sæbø
}
\references{
Sæbø, S., Martens, M. and Martens H. (2010) Three-block data modeling by endo- and
exo-LPLS regression. In Handbook of Partial Least Squares: Concepts, Methods and
Applications. Esposito Vinzi, V.; Chin, W.W.; Henseler, J.; Wang, H. (Eds.). Springer.
}
\keyword{ package }

